Disable proxy when reading userdata with syscfg.
- Created a special resource mgr to disable proxy in a safe was. TODO:
  move to treadmill-core.
