+++ lib/python/treadmill_aws/noproxy.py
@@ -1,0 +1,21 @@
+"""Helper class to disable proxy for given call."""
+
+import os
+
+
+# TODO: this code need to be moved into core.
+class NoProxy():
+    """Safely remove/restore proxy environment variables."""
+
+    def __init__(self):
+        self.saved = dict()
+
+    def __enter__(self):
+        for key in list(os.environ.keys()):
+            if key.lower().endswith('_proxy'):
+                self.saved[key] = os.environ[key]
+                del os.environ[key]
+
+    def __exit__(self, *args, **kwargs):
+        for key, value in self.saved.items():
+            os.environ[key] = value
